689 combine useAppContext and AppProvider files. Remove isAuthenticated bool for authState interface. Set authState object to null if user is logged out rather than use isAuthenticated bool.
